Replaced request with Laravel Data & Added Policy

+++ app/Http/Controllers/Admin/CategoryTypeController.php
@@ -3,20 +3,14 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use BalajiDharma\LaravelAdminCore\Requests\StoreCategoryTypeRequest;
-use BalajiDharma\LaravelAdminCore\Requests\UpdateCategoryTypeRequest;
+use BalajiDharma\LaravelAdminCore\Actions\CategoryType\CategoryTypeCreateAction;
+use BalajiDharma\LaravelAdminCore\Actions\CategoryType\CategoryTypeUpdateAction;
+use BalajiDharma\LaravelAdminCore\Data\CategoryType\CategoryTypeCreateData;
+use BalajiDharma\LaravelAdminCore\Data\CategoryType\CategoryTypeUpdateData;
 use BalajiDharma\LaravelCategory\Models\CategoryType;
 
 class CategoryTypeController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('can:category.type list', ['only' => ['index']]);
-        $this->middleware('can:category.type create', ['only' => ['create', 'store']]);
-        $this->middleware('can:category.type edit', ['only' => ['edit', 'update']]);
-        $this->middleware('can:category.type delete', ['only' => ['destroy']]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -24,6 +18,7 @@ class CategoryTypeController extends Controller
      */
     public function index()
     {
+        $this->authorize('adminViewAny', CategoryType::class);
         $categoryTypes = (new CategoryType)->newQuery();
 
         if (request()->has('search')) {
@@ -55,6 +50,8 @@ class CategoryTypeController extends Controller
      */
     public function create()
     {
+        $this->authorize('adminCreate', CategoryType::class);
+
         return view('admin.category.type.create');
     }
 
@@ -63,18 +60,10 @@ class CategoryTypeController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(StoreCategoryTypeRequest $request)
+    public function store(CategoryTypeCreateData $data, CategoryTypeCreateAction $categoryTypeCreateAction)
     {
-        if (! $request->has('is_flat')) {
-            $request['is_flat'] = false;
-        }
-
-        CategoryType::create([
-            'name' => $request->name,
-            'machine_name' => $request->machine_name,
-            'description' => $request->description,
-            'is_flat' => $request->is_flat,
-        ]);
+        $this->authorize('adminCreate', CategoryType::class);
+        $categoryTypeCreateAction->handle($data);
 
         return redirect()->route('admin.category.type.index')
             ->with('message', 'Category type created successfully.');
@@ -88,6 +77,8 @@ class CategoryTypeController extends Controller
      */
     public function edit(CategoryType $type)
     {
+        $this->authorize('adminUpdate', $type);
+
         return view('admin.category.type.edit', compact('type'));
     }
 
@@ -97,13 +88,10 @@ class CategoryTypeController extends Controller
      * @param  \BalajiDharma\LaravelCategory\Models\CategoryType  $categoryType
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(UpdateCategoryTypeRequest $request, CategoryType $type)
+    public function update(CategoryTypeUpdateData $data, CategoryType $type, CategoryTypeUpdateAction $categoryTypeUpdateAction)
     {
-        if (! $request->has('is_flat')) {
-            $request['is_flat'] = false;
-        }
-
-        $type->update($request->all());
+        $this->authorize('adminUpdate', $type);
+        $categoryTypeUpdateAction->handle($data, $type);
 
         return redirect()->route('admin.category.type.index')
             ->with('message', 'Category type updated successfully.');
@@ -117,6 +105,7 @@ class CategoryTypeController extends Controller
      */
     public function destroy(CategoryType $type)
     {
+        $this->authorize('adminDelete', $type);
         $type->delete();
 
         return redirect()->route('admin.category.type.index')
