+++ app/Http/Controllers/Admin/PermissionController.php
@@ -4,19 +4,13 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Permission;
-use BalajiDharma\LaravelAdminCore\Requests\StorePermissionRequest;
-use BalajiDharma\LaravelAdminCore\Requests\UpdatePermissionRequest;
+use BalajiDharma\LaravelAdminCore\Actions\Permission\PermissionCreateAction;
+use BalajiDharma\LaravelAdminCore\Actions\Permission\PermissionUpdateAction;
+use BalajiDharma\LaravelAdminCore\Data\Permission\PermissionCreateData;
+use BalajiDharma\LaravelAdminCore\Data\Permission\PermissionUpdateData;
 
 class PermissionController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('can:permission list', ['only' => ['index', 'show']]);
-        $this->middleware('can:permission create', ['only' => ['create', 'store']]);
-        $this->middleware('can:permission edit', ['only' => ['edit', 'update']]);
-        $this->middleware('can:permission delete', ['only' => ['destroy']]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -24,6 +18,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        $this->authorize('adminViewAny', Permission::class);
         $permissions = (new Permission)->newQuery();
 
         if (request()->has('search')) {
@@ -55,6 +50,8 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        $this->authorize('adminCreate', Permission::class);
+
         return view('admin.permission.create');
     }
 
@@ -63,9 +60,10 @@ class PermissionController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(StorePermissionRequest $request)
+    public function store(PermissionCreateData $data, PermissionCreateAction $permissionCreateAction)
     {
-        Permission::create($request->all());
+        $this->authorize('adminCreate', Permission::class);
+        $permissionCreateAction->handle($data);
 
         return redirect()->route('admin.permission.index')
             ->with('message', __('Permission created successfully.'));
@@ -78,6 +76,8 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
+        $this->authorize('adminView', $permission);
+
         return view('admin.permission.show', compact('permission'));
     }
 
@@ -88,6 +88,8 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        $this->authorize('adminUpdate', $permission);
+
         return view('admin.permission.edit', compact('permission'));
     }
 
@@ -96,9 +98,10 @@ class PermissionController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(UpdatePermissionRequest $request, Permission $permission)
+    public function update(PermissionUpdateData $data, Permission $permission, PermissionUpdateAction $permissionUpdateAction)
     {
-        $permission->update($request->all());
+        $this->authorize('adminUpdate', $permission);
+        $permissionUpdateAction->handle($data, $permission);
 
         return redirect()->route('admin.permission.index')
             ->with('message', __('Permission updated successfully.'));
@@ -111,6 +114,7 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        $this->authorize('adminDelete', $permission);
         $permission->delete();
 
         return redirect()->route('admin.permission.index')
