+++ app/Http/Controllers/Admin/UserController.php
@@ -5,24 +5,16 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Role;
 use App\Models\User;
-use BalajiDharma\LaravelAdminCore\Actions\User\CreateUser;
-use BalajiDharma\LaravelAdminCore\Actions\User\UpdateUser;
-use BalajiDharma\LaravelAdminCore\Requests\User\StoreUserRequest;
-use BalajiDharma\LaravelAdminCore\Requests\User\UpdateUserRequest;
+use BalajiDharma\LaravelAdminCore\Actions\User\UserCreateAction;
+use BalajiDharma\LaravelAdminCore\Actions\User\UserUpdateAction;
+use BalajiDharma\LaravelAdminCore\Data\User\UserCreateData;
+use BalajiDharma\LaravelAdminCore\Data\User\UserUpdateData;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rules;
 
 class UserController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('can:user list', ['only' => ['index', 'show']]);
-        $this->middleware('can:user create', ['only' => ['create', 'store']]);
-        $this->middleware('can:user edit', ['only' => ['edit', 'update']]);
-        $this->middleware('can:user delete', ['only' => ['destroy']]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -30,6 +22,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('adminViewAny', User::class);
         $users = (new User)->newQuery();
 
         if (request()->has('search')) {
@@ -61,6 +54,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('adminCreate', User::class);
         $roles = Role::all();
 
         return view('admin.user.create', compact('roles'));
@@ -71,9 +65,10 @@ class UserController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function store(StoreUserRequest $request, CreateUser $createUser)
+    public function store(UserCreateData $data, UserCreateAction $userCreateAction)
     {
-        $createUser->handle($request->getUserData());
+        $this->authorize('adminCreate', User::class);
+        $userCreateAction->handle($data);
 
         return redirect()->route('admin.user.index')
             ->with('message', __('User created successfully.'));
@@ -86,6 +81,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        $this->authorize('adminView', $user);
         $roles = Role::all();
         $userHasRoles = array_column(json_decode($user->roles, true), 'id');
 
@@ -99,6 +95,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('adminUpdate', $user);
         $roles = Role::all();
         $userHasRoles = array_column(json_decode($user->roles, true), 'id');
 
@@ -110,9 +107,12 @@ class UserController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(UpdateUserRequest $request, User $user, UpdateUser $updateUser)
+    public function update(UserUpdateData $data, User $user, UserUpdateAction $userUpdateAction)
     {
-        $updateUser->handle($request->getUserData(), $user);
+        $this->authorize('adminUpdate', $user);
+        $userUpdateAction->handle($data, $user);
+
+        $user->addMediaFromRequest('image')->toMediaCollection('user-images');
 
         return redirect()->route('admin.user.index')
             ->with('message', __('User updated successfully.'));
@@ -125,6 +125,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('adminDelete', $user);
         $user->delete();
 
         return redirect()->route('admin.user.index')
